requirements.txt
# Core dependencies
pydantic-ai==0.0.14
pydantic==2.10.3
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
httpx==0.28.1

# LLM providers
openai==1.57.2
anthropic==0.40.0
google-generativeai==0.8.3

# LiveKit integration
livekit==0.17.4
livekit-agents

# Additional utilities
python-multipart==0.0.12
websockets==13.1
aiofiles==24.1.0
loguru==0.7.2